+++ src/features/settings/map/components/ConfigurationSettingsGroup.tsx
@@ -3,7 +3,7 @@ import { FaDraftingCompass } from "react-icons/fa";
 import { useTranslation } from "react-i18next";
 import { CollapsibleHeader, SelectInput } from "@components";
 import { useMapView } from "@contexts/MapViewContext";
-import { MAP_OPTIONS } from "@features/atlas/map/constants/map";
+import { MAP_CONFIG_OPTIONS } from "@features/settings";
 
 export function ConfigurationSettingsGroup() {
   const { t } = useTranslation("atlas");
@@ -11,6 +11,8 @@ export function ConfigurationSettingsGroup() {
   const {
     projection,
     setProjection,
+    baseColor,
+    setBaseColor,
     borderColor,
     setBorderColor,
     borderWidth,
@@ -28,22 +30,28 @@ export function ConfigurationSettingsGroup() {
       {showMapSettings && (
         <div>
           <SelectInput
-            label={t("mapSettings.mapProjection")}
+            label={t("mapSettings.projection")}
             value={projection}
             onChange={(v) => setProjection(String(v))}
-            options={MAP_OPTIONS.projection}
+            options={MAP_CONFIG_OPTIONS.projection}
+          />
+          <SelectInput
+            label={t("mapSettings.baseColor")}
+            value={baseColor}
+            onChange={(v) => setBaseColor(String(v))}
+            options={MAP_CONFIG_OPTIONS.baseColor}
           />
           <SelectInput
             label={t("mapSettings.borderColor")}
             value={borderColor}
             onChange={(v) => setBorderColor(String(v))}
-            options={MAP_OPTIONS.strokeColor}
+            options={MAP_CONFIG_OPTIONS.strokeColor}
           />
           <SelectInput
             label={t("mapSettings.borderWidth")}
             value={borderWidth}
             onChange={(v) => setBorderWidth(Number(v))}
-            options={MAP_OPTIONS.strokeWidth}
+            options={MAP_CONFIG_OPTIONS.strokeWidth}
           />
         </div>
       )}